Abort set up process when provider is already configured
